rename QuestionDatabase to QuestionManager
